+++ implementacao/LDS/src/main/java/LDS3/LDS/Request/VantagemRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record VantagemRequest(
     @NotNull Long idEmpresa,
-    @NotNull double valor,
+    @NotNull int valor,
     @NotEmpty String descricao,
     @NotEmpty String nome,
     @NotEmpty String base64Imagem){
